# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "open-cam API",
    "description": "open-cam 是视频流分析与监控管理平台。\n\n- **摄像头**：接入局域网 RTSP 流或本地视频文件，YOLO（cuda/mps/cpu 自适应）检测与跟踪。\n- **规则引擎**：区域入侵 / 徘徊逗留 / 人数统计 / 区域人数 / 越线计数，支持生效时段与冷却去抖。\n- **事件**：命中规则即落库并存快照，可经 VLM（OpenAI 兼容接口）异步复核；处置闭环支持关注星标、负责人指派、状态流转与 webhook 通知，全程留痕。\n- **方案包**：行业规则模板包（连锁零售 / 美容美发 / 餐饮 / 快餐），一键安装与应用。\n- **统计**：分时段进出店客流等聚合视图。\n\n交互式调试见 [Swagger UI](/docs)，结构化阅读见 [ReDoc](/redoc)。\n",
    "contact": {
      "name": "open-cam",
      "url": "https://github.com/local/open-cam"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/cameras/{camera_id}/rules": {
      "get": {
        "tags": [
          "rules"
        ],
        "summary": "摄像头的规则列表",
        "operationId": "list_rules_api_cameras__camera_id__rules_get",
        "parameters": [
          {
            "name": "camera_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Camera Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuleOut"
                  },
                  "title": "Response List Rules Api Cameras  Camera Id  Rules Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "创建规则",
        "description": "name 不传时用规则类型中文名兜底；params 结构随类型而定（见 /api/rules/presets）。",
        "operationId": "create_rule_api_cameras__camera_id__rules_post",
        "parameters": [
          {
            "name": "camera_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Camera Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cameras/{camera_id}/rules/{rule_id}": {
      "put": {
        "tags": [
          "rules"
        ],
        "summary": "更新规则",
        "operationId": "update_rule_api_cameras__camera_id__rules__rule_id__put",
        "parameters": [
          {
            "name": "camera_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Camera Id"
            }
          },
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rule Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "rules"
        ],
        "summary": "删除规则",
        "operationId": "delete_rule_api_cameras__camera_id__rules__rule_id__delete",
        "parameters": [
          {
            "name": "camera_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Camera Id"
            }
          },
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rule Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/events/{event_id}/ack": {
      "post": {
        "tags": [
          "events"
        ],
        "summary": "确认事件（ack）",
        "operationId": "ack_event_api_events__event_id__ack_post",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/packs/install": {
      "post": {
        "tags": [
          "packs"
        ],
        "summary": "安装方案包",
        "description": "source 支持本地目录、.zip 文件路径或 http(s) URL。",
        "operationId": "install_api_packs_install_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/packs/install-upload": {
      "post": {
        "tags": [
          "packs"
        ],
        "summary": "上传并安装方案包",
        "description": "上传 .zip 方案包并安装。",
        "operationId": "install_upload_api_packs_install_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_install_upload_api_packs_install_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/packs/{pack_id}/apply": {
      "post": {
        "tags": [
          "packs"
        ],
        "summary": "应用方案包",
        "description": "新包不传 camera_id，按包创建多路摄像头；旧包必须指定 camera_id。",
        "operationId": "apply_api_packs__pack_id__apply_post",
        "parameters": [
          {
            "name": "pack_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pack Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackApplyOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/packs/{pack_id}": {
      "get": {
        "tags": [
          "packs"
        ],
        "summary": "方案包详情",
        "description": "规范化 PackDetail；无效/不兼容包仍返回，availability 标明原因。",
        "operationId": "get_pack_detail_api_packs__pack_id__get",
        "parameters": [
          {
            "name": "pack_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pack Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "packs"
        ],
        "summary": "卸载方案包",
        "description": "仅已安装的包可卸载；内置包返回 400。",
        "operationId": "uninstall_api_packs__pack_id__delete",
        "parameters": [
          {
            "name": "pack_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pack Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/packs/{pack_id}/assets/{asset_id}": {
      "get": {
        "tags": [
          "packs"
        ],
        "summary": "方案包白名单资产",
        "description": "仅已声明且校验通过的资产；支持 Range、MIME、ETag。",
        "operationId": "get_pack_asset_api_packs__pack_id__assets__asset_id__get",
        "parameters": [
          {
            "name": "pack_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pack Id"
            }
          },
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApplyRequest": {
        "properties": {
          "camera_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Camera Id"
          }
        },
        "type": "object",
        "title": "ApplyRequest"
      },
...
      "Body_install_upload_api_packs_install_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_install_upload_api_packs_install_upload_post"
      },
...
      "CameraOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "source_uri": {
            "type": "string",
            "title": "Source Uri"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "health": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CameraHealth"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "source_type",
          "source_uri",
          "status"
        ],
        "title": "CameraOut"
      },
...
      "EventOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "camera_id": {
            "type": "integer",
            "title": "Camera Id"
          },
          "rule_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "ts": {
            "type": "number",
            "title": "Ts"
          },
          "snapshot_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Path"
          },
          "source_offset": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Offset"
          },
          "camera_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Camera Name"
          },
          "source_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Filename"
          },
          "detail": {
            "additionalProperties": true,
            "type": "object",
            "title": "Detail"
          },
          "vlm_status": {
            "type": "string",
            "title": "Vlm Status"
          },
          "vlm_verdict": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vlm Verdict"
          },
          "vlm_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vlm Reason"
          },
          "acked": {
            "type": "boolean",
            "title": "Acked"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "starred": {
            "type": "boolean",
            "title": "Starred"
          },
          "assignee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee"
          },
          "assignee_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee Id"
          },
          "verdict": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verdict"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "intent": {
            "type": "string",
            "title": "Intent",
            "default": "alert"
          },
          "needs_action": {
            "type": "boolean",
            "title": "Needs Action",
            "default": true
          },
          "repeat_count": {
            "type": "integer",
            "title": "Repeat Count",
            "default": 1
          },
          "clip_start": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clip Start",
            "readOnly": true
          },
          "clip_end": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clip End",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "camera_id",
          "rule_id",
          "type",
          "confidence",
          "ts",
          "snapshot_path",
          "detail",
          "vlm_status",
          "vlm_verdict",
          "vlm_reason",
          "acked",
          "status",
          "starred",
          "assignee",
          "note",
          "clip_start",
          "clip_end"
        ],
        "title": "EventOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InstallRequest": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "source"
        ],
        "title": "InstallRequest"
      },
...
      "PackApplyOut": {
        "properties": {
          "cameras": {
            "items": {
              "$ref": "#/components/schemas/CameraOut"
            },
            "type": "array",
            "title": "Cameras"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/RuleOut"
            },
            "type": "array",
            "title": "Rules"
          }
        },
        "type": "object",
        "required": [
          "cameras",
          "rules"
        ],
        "title": "PackApplyOut"
      },
...
      "PackDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "vertical": {
            "type": "string",
            "title": "Vertical"
          },
          "author": {
            "type": "string",
            "title": "Author",
            "default": ""
          },
          "origin": {
            "type": "string",
            "title": "Origin"
          },
          "fingerprint": {
            "type": "string",
            "title": "Fingerprint"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "availability": {
            "type": "string",
            "title": "Availability"
          },
          "unavailable_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unavailable Reason"
          },
          "presentation": {
            "$ref": "#/components/schemas/PackPresentationOut"
          },
          "cameras": {
            "items": {
              "$ref": "#/components/schemas/PackCameraDetailOut"
            },
            "type": "array",
            "title": "Cameras"
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/PackRuleDetailOut"
            },
            "type": "array",
            "title": "Rules"
          },
          "experience": {
            "$ref": "#/components/schemas/PackExperienceOut"
          },
          "application": {
            "$ref": "#/components/schemas/PackApplicationOut"
          },
          "privacy": {
            "$ref": "#/components/schemas/PackPrivacyOut"
          },
          "readme_html": {
            "type": "string",
            "title": "Readme Html",
            "default": ""
          },
          "min_opencam_version": {
            "type": "string",
            "title": "Min Opencam Version",
            "default": "0.1.0"
          },
          "format_version": {
            "type": "integer",
            "title": "Format Version",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "version",
          "vertical",
          "origin",
          "fingerprint",
          "availability",
          "presentation",
          "cameras",
          "rules",
          "experience",
          "application"
        ],
        "title": "PackDetail"
      },
...
      "RuleCreate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "type": {
            "type": "string",
            "pattern": "^(zone_intrusion|loitering|object_count|zone_count|line_crossing)$",
            "title": "Type"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "cooldown": {
            "type": "number",
            "title": "Cooldown",
            "default": 30.0
          },
          "intent": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(observe|alert)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intent"
          },
          "escalate": {
            "additionalProperties": true,
            "type": "object",
            "title": "Escalate"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "RuleCreate"
      },
      "RuleOut": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "type": {
            "type": "string",
            "pattern": "^(zone_intrusion|loitering|object_count|zone_count|line_crossing)$",
            "title": "Type"
          },
          "params": {
            "additionalProperties": true,
            "type": "object",
            "title": "Params"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "cooldown": {
            "type": "number",
            "title": "Cooldown",
            "default": 30.0
          },
          "intent": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(observe|alert)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intent"
          },
          "escalate": {
            "additionalProperties": true,
            "type": "object",
            "title": "Escalate"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "camera_id": {
            "type": "integer",
            "title": "Camera Id"
          }
        },
        "type": "object",
        "required": [
          "type",
          "id",
          "camera_id"
        ],
        "title": "RuleOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}